Fix casting issue with MAX_BUFFER_SIZE.
cpl_vsil_buffered_reader.cpp:262:45: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]


git-svn-id: https://svn.osgeo.org/gdal/trunk@34915 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsil_buffered_reader.cpp
@@ -259,7 +259,8 @@ size_t VSIBufferedReaderHandle::Read( void *pBuffer, size_t nSize, size_t nMemb 
                     1, nToReadInFile);
             const size_t nRead = nReadInBuffer + nReadInFile;
 
-            nBufferSize = static_cast<int>( MIN(nRead, MAX_BUFFER_SIZE) );
+            nBufferSize = static_cast<int>(
+                MIN(nRead, static_cast<size_t>(MAX_BUFFER_SIZE)) );
             nBufferOffset = nCurOffset + nRead - nBufferSize;
             memcpy(pabyBuffer,
                    static_cast<GByte *>(pBuffer) + nRead - nBufferSize,
@@ -290,7 +291,8 @@ size_t VSIBufferedReaderHandle::Read( void *pBuffer, size_t nSize, size_t nMemb 
             return 0;
         bNeedBaseHandleSeek = false;
         const size_t nReadInFile = m_poBaseHandle->Read(pBuffer, 1, nTotalToRead);
-        nBufferSize = static_cast<int>( MIN(nReadInFile, MAX_BUFFER_SIZE) );
+        nBufferSize = static_cast<int>(
+            MIN(nReadInFile, static_cast<size_t>(MAX_BUFFER_SIZE)) );
         nBufferOffset = nCurOffset + nReadInFile - nBufferSize;
         memcpy(pabyBuffer, (GByte*)pBuffer + nReadInFile - nBufferSize, nBufferSize);
